Bootstrapped Template and Resources for Splash Page
Got a template from styleshout and componentized the sections of their splash page. Replacing entire site with the splash page right now till we get auth in. Included some Jquery code for now for styling purposes. We should Reactify them  as much as we can and get rid of Jquery